local.requirements.txt
# Requirements for discord botte bot

setuptools
PyNaCl
opuslib
nextcord
python-dotenv
humanfriendly
paramiko
blagues-api
